Expose hex center through draw data. Add new direction array for clients that render with Y increasing upward. Encapsulate neighbor relationships in the hexagon group.

+++ Sixsmith/Sixsmith/Directions.swift
@@ -1,7 +1,6 @@
 
 public struct Direction {
-    
-    let value: UInt8
+    let value: Int
     
     static let zero = Direction(value: 0)
     static let one = Direction(value:1)
